backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Cors==4.0.0
python-dotenv==1.0.0
Werkzeug==2.3.7
cryptography==41.0.3
twilio==8.10.0
celery==5.3.1
redis==5.0.0
google-generativeai==0.3.0
requests==2.31.0
